test(projects): keep locked policy proof bounded
